extract unvalidated class file information into submodules
